refactor(files): 🔨 rename files from CamelCase to kebab-case
Renamed files across the project to adhere to kebab-case naming convention.
Example: exampleContext.ts → example-context.ts
