+++ apps/nextjs/src/common/contexts/example-context.tsx
