+++ apps/nextjs/src/common/contexts/index.ts
@@ -1,1 +1,1 @@
-export * from './ExampleContext';
+export * from './example-context';
