+++ apps/nextjs/src/common/hooks/use-example.ts
